requirements.txt
autopep8
Django
django-cors-headers
django-filter
djangorestframework
drf-yasg
swagger-ui-py
djangorestframework-simplejwt
Pillow
